Procedure Unfinished
My aim for this procedure was to pretty print funds within the range, so
I thought to use a cursor. I'm trying to get a table that has one tuple
for each fund with the latest closing price within the range, ordered by
price.

Is there an easier way to go about doing this? If not, how should this
be accomplished?